+++ NEWS.md
@@ -6,6 +6,10 @@
 
 * Improve documentation of interval statistics with more complex examples.
 
+## Bug fixes
+
+* `bed_merge()` now reports start/end columns if spec is provided (#288)
+
 # valr 0.3.0
 
 ## Enhancements